backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ML/Embeddings
sentence-transformers>=2.2.0
openai>=1.0.0

# Vector Database
chromadb>=0.5.0

# Data Processing (using versions with prebuilt wheels)
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Dimensionality Reduction (for UMAP later)
umap-learn>=0.5.0

# Utilities
tqdm==4.66.1
